[Fix] MIDI export: Little endian handling fix (not sure if it affects anything). [Fix] VST: VST host version fields didn't follow the actual version of OpenMPT (no behaviour changes expected).
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@938 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/mod2midi.h
@@ -4,7 +4,7 @@
 typedef struct _MOD2MIDIINSTR
 {
 	UINT nChannel;
-	UINT nProgram;
+	uint8 nProgram;
 } MOD2MIDIINSTR, *PMOD2MIDIINSTR;
